Bug "fix" - Combo showing up huge and not following rotations
The "fix" for this bug is to just not support 360 environments 🤣
So I've implemented a check for the normal environment type and I cancel the initialisation of the counter. It's not really pretty, but I don't think too many people would care anyway. I can always properly fix it if I get people asking me to support other environment types.